+++ dinky-admin/src/main/java/org/dinky/service/resource/BaseResourceManager.java
@@ -21,8 +21,8 @@ package org.dinky.service.resource;
 
 import org.dinky.data.model.SystemConfiguration;
 import org.dinky.service.resource.impl.HdfsResourceManager;
-
 import org.dinky.service.resource.impl.OssResourceManager;
+
 import org.springframework.web.multipart.MultipartFile;
 
 import cn.hutool.core.io.FileUtil;
@@ -51,6 +51,7 @@ public interface BaseResourceManager {
     }
 
     default String getFile(String path) {
-        return FileUtil.normalize(FileUtil.file(instances.getResourcesUploadBasePath().getValue(), path).toString());
+        return FileUtil.normalize(
+                FileUtil.file(instances.getResourcesUploadBasePath().getValue(), path).toString());
     }
 }
